requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --extra-index-url=https://packagecloud.io/cytomine-uliege/Cytomine-python-client/pypi/simple
#
--extra-index-url https://packagecloud.io/cytomine-uliege/Cytomine-python-client/pypi/simple

affine==2.4.0
    # via rasterio
aiofiles==23.1.0
    # via cytomine-pims (setup.py)
aioredis[hiredis]==2.0.1
    # via cytomine-pims (setup.py)
amqp==5.1.1
    # via kombu
anyio==3.6.2
    # via watchfiles
asciitree==0.3.3
    # via zarr
asgiref==3.6.0
    # via cytomine-pims (setup.py)
async-timeout==4.0.2
    # via aioredis
attrs==22.2.0
    # via rasterio
billiard==3.6.4.0
    # via celery
cachecontrol==0.12.11
    # via cytomine-python-client
celery==5.2.7
    # via cytomine-pims (setup.py)
certifi==2022.12.7
    # via
    #   rasterio
    #   requests
cffi==1.15.1
    # via pyvips
charset-normalizer==3.1.0
    # via requests
click==8.1.3
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   cligj
    #   rasterio
    #   uvicorn
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via
    #   celery
    #   rasterio
click-repl==0.2.0
    # via celery
cligj==0.7.2
    # via rasterio
contourpy==1.0.7
    # via matplotlib
cycler==0.11.0
    # via matplotlib
cytomine-python-client==2.9.0
    # via cytomine-pims (setup.py)
entrypoints==0.4
    # via numcodecs
fastapi==0.68.2
    # via cytomine-pims (setup.py)
fasteners==0.18
    # via zarr
fonttools==4.38.0
    # via matplotlib
future==0.18.3
    # via cytomine-python-client
h11==0.14.0
    # via uvicorn
hiredis==2.2.2
    # via aioredis
httptools==0.5.0
    # via uvicorn
idna==3.4
    # via
    #   anyio
    #   requests
imagecodecs==2023.1.23
    # via cytomine-pims (setup.py)
imageio==2.26.0
    # via scikit-image
importlib-metadata==6.0.0
    # via cytomine-pims (setup.py)
kiwisolver==1.4.4
    # via matplotlib
kombu==5.2.4
    # via celery
lazy-loader==0.1
    # via scikit-image
markdown-it-py==2.2.0
    # via rich
matplotlib==3.7.1
    # via cytomine-pims (setup.py)
mdurl==0.1.2
    # via markdown-it-py
msgpack==1.0.4
    # via cachecontrol
networkx==3.0
    # via scikit-image
numcodecs==0.11.0
    # via zarr
numpy==1.24.2
    # via
    #   contourpy
    #   cytomine-pims (setup.py)
    #   cytomine-python-client
    #   imagecodecs
    #   imageio
    #   matplotlib
    #   numcodecs
    #   opencv-python-headless
    #   pywavelets
    #   rasterio
    #   scikit-image
    #   scipy
    #   shapely
    #   snuggs
    #   tifffile
    #   zarr
opencv-python-headless==4.7.0.72
    # via cytomine-python-client
orjson==3.8.7
    # via cytomine-pims (setup.py)
packaging==23.0
    # via
    #   matplotlib
    #   scikit-image
pathvalidate==2.5.2
    # via cytomine-pims (setup.py)
pillow==9.4.0
    # via
    #   cytomine-pims (setup.py)
    #   cytomine-python-client
    #   imageio
    #   matplotlib
    #   scikit-image
pint==0.20.1
    # via cytomine-pims (setup.py)
pkgconfig==1.5.5
    # via pyvips
prompt-toolkit==3.0.38
    # via click-repl
pycparser==2.21
    # via cffi
pydantic==1.10.5
    # via
    #   cytomine-pims (setup.py)
    #   fastapi
pydicom==2.3.1
    # via cytomine-pims (setup.py)
pygments==2.14.0
    # via rich
pyparsing==3.0.9
    # via
    #   matplotlib
    #   snuggs
python-dateutil==2.8.2
    # via
    #   cytomine-pims (setup.py)
    #   matplotlib
python-dotenv==1.0.0
    # via
    #   cytomine-pims (setup.py)
    #   uvicorn
python-gdcm==3.0.21
    # via cytomine-pims (setup.py)
python-multipart==0.0.6
    # via cytomine-pims (setup.py)
pytz==2022.7.1
    # via celery
pyvips==2.2.1
    # via cytomine-pims (setup.py)
pywavelets==1.4.1
    # via scikit-image
pyyaml==6.0
    # via uvicorn
rasterio==1.3.6
    # via cytomine-pims (setup.py)
requests==2.28.2
    # via
    #   cachecontrol
    #   cytomine-python-client
    #   requests-toolbelt
requests-toolbelt==0.10.1
    # via cytomine-python-client
rich==13.3.2
    # via cytomine-pims (setup.py)
scikit-image==0.20.0
    # via cytomine-pims (setup.py)
scipy==1.9.1
    # via scikit-image
shapely==2.0.1
    # via
    #   cytomine-pims (setup.py)
    #   cytomine-python-client
six==1.16.0
    # via
    #   click-repl
    #   cytomine-python-client
    #   python-dateutil
sniffio==1.3.0
    # via anyio
snuggs==1.4.7
    # via rasterio
starlette==0.14.2
    # via fastapi
tifffile==2023.2.28
    # via
    #   cytomine-pims (setup.py)
    #   scikit-image
typing-extensions==4.5.0
    # via
    #   aioredis
    #   pydantic
urllib3==1.26.14
    # via
    #   cytomine-python-client
    #   requests
uvicorn[standard]==0.20.0
    # via cytomine-pims (setup.py)
uvloop==0.17.0
    # via uvicorn
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
watchfiles==0.18.1
    # via uvicorn
wcwidth==0.2.6
    # via prompt-toolkit
websockets==10.4
    # via uvicorn
zarr==2.14.2
    # via cytomine-pims (setup.py)
zipp==3.15.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
